// src/api/types.gen.ts
export type GetApiTablesByIdResponses = {
    /**
     * Full table data
     */
    200: {
        id: number;
        name: string;
        createdAt?: string | null;
        columns: Array<{
            id: number;
            name: string;
            type: 'text' | 'number' | 'select' | 'toggle' | 'date';
            visible: boolean;
            options: Array<{
                id: number;
                label: string;
                color: string | null;
                order: number;
                columnId: number;
            }>;
        }>;
        rows: Array<{
            id: number;
            order: number;
            createdAt?: string | null;
            values: Array<{
                id: number;
                rowId: number;
                columnId: number;
                value: string | null;
            }>;
        }>;
    };
};